refactor: renomeia nome dos metodos na service e controller
